+++ src/main/java/com/spring_boots/spring_boots/user/controller/UserAdminApiController.java
@@ -5,17 +5,14 @@ import com.spring_boots.spring_boots.user.dto.UserDto;
 import com.spring_boots.spring_boots.user.dto.request.AdminCodeRequestDto;
 import com.spring_boots.spring_boots.user.dto.request.AdminGrantTokenRequestDto;
 import com.spring_boots.spring_boots.user.dto.response.*;
+import com.spring_boots.spring_boots.user.service.AdminService;
 import com.spring_boots.spring_boots.user.service.UserService;
 import lombok.RequiredArgsConstructor;
-import org.apache.coyote.Response;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 import static com.spring_boots.spring_boots.config.jwt.UserConstants.ACCESS_TOKEN_TYPE_VALUE;
 
@@ -24,6 +21,7 @@ import static com.spring_boots.spring_boots.config.jwt.UserConstants.ACCESS_TOKE
 @RequiredArgsConstructor
 public class UserAdminApiController {
 
+    private final AdminService adminService;
     private final UserService userService;
 
     //모든 회원 정보 조회(관리자)
@@ -34,15 +32,15 @@ public class UserAdminApiController {
             @RequestParam(defaultValue = "10") int size
 //            ,@RequestParam(value = "keyword", defaultValue = "") String keyword
     ) {
-//        List<UserResponseDto> users = userService.findAll();
-        Page<UserResponseDto> result = userService.getUsersByCreatedAt(page, size);
+        Page<UserResponseDto> result = adminService.getUsersByCreatedAt(page, size);
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
 
+    //회원수 세기
     @PreAuthorize("hasRole('ADMIN')")
     @GetMapping("/admin/count")
     public ResponseEntity<UserAdminCountResponseDto> countUsers() {
-        UserAdminCountResponseDto userAdminCountResponseDto = userService.countUsers();
+        UserAdminCountResponseDto userAdminCountResponseDto = adminService.countUsers();
         return ResponseEntity.status(HttpStatus.OK).body(userAdminCountResponseDto);
     }
 
@@ -75,7 +73,7 @@ public class UserAdminApiController {
                     .body(AdminGrantTokenResponseDto.builder()
                             .message("인증되지 않은 사용자입니다. 로그인해주세요").build());
         }
-        userService.grantRole(authUser, adminGrantTokenRequestDto);
+        adminService.grantRole(authUser, adminGrantTokenRequestDto);
         return ResponseEntity.status(HttpStatus.OK).body(AdminGrantTokenResponseDto.builder()
                 .message("권한부여 성공!").build());
     }
@@ -92,7 +90,7 @@ public class UserAdminApiController {
         }
 
         try {
-            boolean isAdmin = userService.validateAdminToken(accessToken);
+            boolean isAdmin = adminService.validateAdminToken(accessToken);
 
             if (isAdmin) {
                 return ResponseEntity.status(HttpStatus.OK)
@@ -111,7 +109,7 @@ public class UserAdminApiController {
     @PreAuthorize("hasRole('ADMIN')")
     @PostMapping("/users/grant")
     public ResponseEntity<AdminCodeResponseDto> checkAdminCode(@RequestBody AdminCodeRequestDto adminCodeDto) {
-        if (userService.checkAdminCode(adminCodeDto)) {
+        if (adminService.checkAdminCode(adminCodeDto)) {
             return ResponseEntity.status(HttpStatus.OK)
                     .body(AdminCodeResponseDto.builder().message("success").build());
         } else {
