+++ src/main/java/com/spring_boots/spring_boots/user/service/UserService.java
@@ -172,10 +172,6 @@ public class UserService {
         return userRepository.existsByUserRealId(userRealId);
     }
 
-    @Transactional
-    public void grantRole(Users authUser, AdminGrantTokenRequestDto adminGrantTokenRequestDto) {
-        authUser.updateToRole(adminGrantTokenRequestDto);
-    }
 
     public boolean isGrantAdmin(Users authUser) {
         return authUser.getRole().equals(UserRole.ADMIN);
@@ -185,22 +181,7 @@ public class UserService {
         return jwtProvider.validateToken(accessToken);
     }
 
-    public boolean validateAdminToken(String accessToken) {
-        return jwtProvider.validateAdminToken(accessToken);
-    }
 
-    //관리자코드체크
-    public boolean checkAdminCode(AdminCodeRequestDto adminCodeDto) {
-        //임의 토큰 만들기
-        String tempAdminCode = bCryptPasswordEncoder.encode(adminCode);
-        String adminCode = adminCodeDto.getAdminCode();
-        if (bCryptPasswordEncoder.matches(adminCode, tempAdminCode)) {
-            return true;
-        } else {
-            log.info("잘못된 관리자 토큰");
-            return false;
-        }
-    }
 
     //엔티티 변경
     public Users getUserEntityByDto(UserDto userDto) {
@@ -234,26 +215,6 @@ public class UserService {
         return user.isDeleted();
     }
 
-    public Page<UserResponseDto> getUsersByCreatedAt(int page, int size) {
-        Pageable pageable = PageRequest.of(page, size);
-        Page<Users> usersPage = userRepository.findAll(pageable);
-
-        return usersPage.map(Users::toResponseDto);
-    }
-
-    public UserAdminCountResponseDto countUsers() {
-        List<Users> users = userRepository.findAll();
-        long countAdmin = users.stream()
-                .filter(user -> user.getRole().equals(UserRole.ADMIN))
-                .count();
-        long totalUsers = users.stream()
-                .filter(user -> !user.isDeleted())
-                .count();
-        return UserAdminCountResponseDto.builder()
-                .countAdmin(countAdmin)
-                .totalUser(totalUsers)
-                .build();
-    }
 
     public boolean validateSignup(UserSignupRequestDto userSignupRequestDto) {
         // username 유효성 검증: 2~20글자, 숫자 포함 불가
